Split types into their own files
